Revert " feat：修改默认构建配置 & subpackages分包模版"
