--- .env.development
@@ -1,7 +1,0 @@
-# 配置要求：必须以TARO_APP_开头否则不可以通过process.env.TARO_APP_XX获得
-
-# 域名
-TARO_APP_DOMAIN="https://hxtia.com:80"
-
-# 请求根目录
-TARO_APP_BASE_ROOT_PATH="/api"
