+++ config/prod.js
@@ -2,39 +2,36 @@ module.exports = {
   env: {
     NODE_ENV: '"production"'
   },
-  defineConstants: {},
-  mini: {
-    /** 自动提取了chunk */
-    // commonChunks: {},
+  defineConstants: {
+  },
+  mini: {},
+  h5: {
+    /**
+     * WebpackChain 插件配置
+     * @docs https://github.com/neutrinojs/webpack-chain
+     */
+    // webpackChain (chain) {
+    //   /**
+    //    * 如果 h5 端编译后体积过大，可以使用 webpack-bundle-analyzer 插件对打包体积进行分析。
+    //    * @docs https://github.com/webpack-contrib/webpack-bundle-analyzer
+    //    */
+    //   chain.plugin('analyzer')
+    //     .use(require('webpack-bundle-analyzer').BundleAnalyzerPlugin, [])
 
-    /** 优化主包体积 & 自动提取chunk */
-    optimizeMainPackage: {
-      enable: true,
-    },
-
-    /** 提取css */
-    miniCssExtractPluginOption: {
-      filename: '[name].css',
-      chunkFilename: '[name].css',
-    },
+    //   /**
+    //    * 如果 h5 端首屏加载时间过长，可以使用 prerender-spa-plugin 插件预加载首页。
+    //    * @docs https://github.com/chrisvfritz/prerender-spa-plugin
+    //    */
+    //   const path = require('path')
+    //   const Prerender = require('prerender-spa-plugin')
+    //   const staticDir = path.join(__dirname, '..', 'dist')
+    //   chain
+    //     .plugin('prerender')
+    //     .use(new Prerender({
+    //       staticDir,
+    //       routes: [ '/pages/index/index' ],
+    //       postProcess: (context) => ({ ...context, outputPath: path.join(staticDir, 'index.html') })
+    //     }))
+    // }
   }
-  /** 看起来会默认压缩 */
-  // jsMinimizer: 'terser', // 默认压缩为terser
-  // terser: {
-  //   enable: true,
-  //   // https://github.com/terser/terser#minify-options
-  //   config: {
-  //     compress: {
-  //       arguments: true,
-  //       dead_code: true
-  //     },
-  //     toplevel: true,
-  //     keep_classnames: true,
-  //     keep_fnames: true
-  //   },
-  // },
-  // cssMinimize: {}
-  // plugins: [
-  //   '@tarojs/plugin-terser'
-  // ],
 }
