+++ package.json
@@ -70,7 +70,6 @@
     "@babel/core": "^7.8.0",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.5",
     "@tarojs/cli": "3.5.11",
-    "@tarojs/plugin-terser": "^2.2.10",
     "@tarojs/webpack5-runner": "3.5.11",
     "@types/react": "^18.0.0",
     "@types/react-redux": "^7.1.33",
@@ -79,17 +78,17 @@
     "@typescript-eslint/parser": "^5.20.0",
     "babel-plugin-import": "^1.13.6",
     "babel-preset-taro": "3.5.11",
-    "chalk": "^3.0.0",
     "eslint": "^8.12.0",
     "eslint-config-taro": "3.5.11",
     "eslint-plugin-import": "^2.12.0",
     "eslint-plugin-react": "^7.8.2",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "exec-sh": "^0.4.0",
     "react-refresh": "^0.11.0",
     "stylelint": "^14.4.0",
     "typescript": "^4.1.0",
-    "webpack": "5.69.0"
+    "webpack": "5.69.0",
+    "chalk": "^3.0.0",
+    "exec-sh": "^0.4.0"
   },
   "resolutions": {
     "@swc/core": "1.3.6"
