+++ src/components/Error/index.module.scss
@@ -1,6 +1,6 @@
-@import 'src/app.scss';
+@import '../../app.scss';
 
-.ForbiddenContainer {
+.ErrorContainer {
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -12,6 +12,7 @@
     position: relative;
     display: flex;
     flex-direction: column;
+    justify-content: space-around;
     width: 80%;
     height: 70%;
 
@@ -28,7 +29,6 @@
       text-align: center;
       font-size: 36px;
       font-weight: 700;
-      margin-top: 10%;
       color: $base-element-color;
     }
   }
