+++ src/components/Welcome/index.tsx
@@ -1,0 +1,7 @@
+import { View } from '@tarojs/components'
+
+const Index = () => {
+  return <View>欢迎你</View>
+}
+
+export default Index
