--- src/components/common/README.md
@@ -1,1 +1,0 @@
-# 存放业务基础组件
