+++ src/pages/Index/index.config.ts
