+++ tsconfig.json
@@ -22,12 +22,7 @@
     "typeRoots": [
       "node_modules/@types"
     ],
-    "jsxImportSource": "react",
-    "paths": {
-      "src/*": [
-        "./src/*"
-      ]
-    }
+    "jsxImportSource": "react"
   },
   "include": [
     "./src",
